Add support for creating directories
Motivation:

The UPLOAD activity in a macaroon, in combination with a Path caveat,
can either allow a user to upload a single file or upload arbitrary
files into a directory with the path as a prefix.  The allowed behaviour
depends on whether the path already exists in dCache and is a directory.

Therefore, in order to support uploading arbitrary files into a
directory, the UPLOAD activity and Path may be used, but the target
directory must already exist.

Modification:

This patch adds a simple WebDAV client that can create a directory.

Some substantial refactoring was necessary, so that the client may be
shared between the macaroon generator and the WebDAV client.

As part of this refactoring, the authentication was moved from manually
adding a header per-request, to being included as part of the client.

Result:

It is now possible for a macaroon to allow the client to upload
arbitrary files into a directory.