Avoid requesting the 'path' caveat twice.
Motivation:

Currently, the dynamic agent requests a macaroon with the "path" caveat
appearing twice: once implicitly from the POST request URL's path and
once as an explicit caveat.

As "path" caveats combine by concatination (even when absolute) the
result of having the path caveat twice is a macaroon that cannot be used
to target the desired file.

Modification:

Remove the explicit "path" caveat in the macaroon request.

Result:

A valid macaroon that targets the desired file.